Enhance AssignmentEditor with allowedRoleIds prop
Added allowedRoleIds prop to AssignmentEditor component and updated role selection logic to filter based on allowed roles.
This is to allow selection of eligible roles for the relationship. The roles that were not originally requested for the relationship are grayed out and not selectable.
Also when using templates for roles, if template contains role that is not eligible, it is noted and warned.